rename to HyperdriveOffsetComparisonService, add tests
